requirements.txt
pytesseract
pdf2image
openai
python-dotenv
pandas
streamlit
